+++ src/components/OptionsList/OptionsList.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../test/jest/__mock__';
@@ -67,6 +68,12 @@ const renderOptionsList = ({
 describe('OptionsList', () => {
   afterAll(() => {
     onSelectMock.mockClear();
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderOptionsList(notEmptyOptionsListProps);
+
+    await runAxeTest({ rootNode: container });
   });
 
   describe('when there are options', () => {