+++ src/components/ProfileTree/ProfileLinker/LinkerTrigger.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../../test/jest/__mock__';
@@ -37,6 +38,12 @@ const renderLinkerTrigger = ({
 };
 
 describe('LinkerTrigger', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderLinkerTrigger({ ...linkerTriggerProps });
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should be rendered with node type title', () => {
     const { getByText } = renderLinkerTrigger({
       ...linkerTriggerProps,