+++ src/components/Spinner/Spinner.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
@@ -29,6 +30,12 @@ const renderSpinner = ({ entity }) => {
 };
 
 describe('Spinner component', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderSpinner(spinnerProps);
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should render preloader', () => {
     const { getByText } = renderSpinner(spinnerProps);