+++ src/settings/JobProfiles/tests/JobProfilesForm.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { fireEvent } from '@testing-library/react';
-
 import { noop } from 'lodash';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import { buildResources } from '@folio/stripes-data-transfer-components/test/helpers';
@@ -135,6 +135,12 @@ describe('<JobProfilesForm>', () => {
 
   afterAll(() => {
     delete global.fetch;
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderJobProfilesForm(jobProfilesFormProps({}));
+
+    await runAxeTest({ rootNode: container });
   });
 
   describe('when profile is duplicated', () => {