+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/tests/InvoiceInformation.test.js
@@ -5,6 +5,7 @@ import {
 } from '@testing-library/react';
 import { noop } from 'lodash';
 
+import { runAxeTest } from '@folio/stripes-testing';
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
 import '../../../../../../../test/jest/__mock__';
@@ -27,6 +28,12 @@ const renderInvoiceInformation = () => {
 };
 
 describe('InvoiceInformation', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderInvoiceInformation();
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should render correct fields', () => {
     const { getByText } = renderInvoiceInformation();